requirements.txt
pandas>=2.0.0
streamlit>=1.28.0
gspread>=5.10.0
oauth2client>=4.1.3
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
pdfplumber>=0.10.0
groq>=0.4.0
openai>=1.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0
python-dotenv>=1.0.0
torch>=2.0.0